Membership and "Hours per Day" support.
Added membership support and index for projects.
Added a method to check membership on a project.
Updated the project-edit page for setting memberships.
Updated the ProjectController to handle new membership fields.
Updated the project-edit page for setting number of hours per day.
Updated the ProjectController to handle the hours per day fields.