Adding support for "detaching" a connection. This will allow properly shutting down the connections in the near future.
Found an interesting situation in the destruction of a Connection. The
connection tries to delete the thread while the thread is trying to shutdown
the connection.
See Connection_secure.cpp:60
See Thread.cpp:50
See Thread.cpp:69 (Thread::abort)
See http://pubs.opengroup.org/onlinepubs/7908799/xsh/pthread_cancel.html

It looks like the thread just needs to be marked as not running before it
gets to the point where it tries to delete the thread object. Maybe mark the
work as completed BEFORE invoking the client cleanup code?