Adding a unit test to check that corrupted encrypted data fails to decode.
This did require adding a friend attribute on the document object. Not sure
how I feel about that.  Might need to find a better way to give unit tests the
ability to violate the invariant nature of classes.  Not loving the design of
my test runner framework currently. Something seems very frail about it.