Ground work for the gnutls integration.
I played around a bit with the gnutlsxx header.  Doesn't seem to support all of the functions
used in the example code. So I think I am going to go ahead and roll my own C++ wrapper that
uses more C++11 stuff, and makes more sense for the logjam and logjamd division.